feat: extract PublicLayout component + Tauri desktop padding for public pages
Refactor duplicated sticky nav and layout wrapper across public pages
into a shared <PublicLayout> component. Add data-public-layout attribute
and corresponding CSS rule for Tauri desktop padding-top support.

- Create src/web/src/components/public-layout.tsx with configurable
  maxWidth, breadcrumb, rightSlot, and footer props
- Migrate templates, templates/[id], privacy, and blog pages to use it
- Add data-public-layout to studio/new root divs for Tauri padding
- Add CSS: html[data-tauri-desktop] [data-public-layout] { padding-top: 1.75rem }
