+++ src/web/src/app/templates/[id]/client.tsx
@@ -1,9 +1,9 @@
 "use client";
 
 import Link from "next/link";
-import Image from "next/image";
 import { ArrowLeft, Check } from "lucide-react";
 import { buttonVariants } from "@/components/ui/button";
+import { PublicLayout } from "@/components/public-layout";
 import { MemberCard } from "./_components/member-card";
 import type { TemplatePreset } from "@/lib/templates";
 
@@ -31,29 +31,21 @@ export function TemplateDetailClient({
     : `/sign-in?redirect=${encodeURIComponent(getUrl)}`;
 
   return (
-    <div className="min-h-dvh bg-background">
-      {/* Nav */}
-      <nav className="sticky top-0 z-50 border-b bg-background/90 backdrop-blur-sm">
-        <div className="mx-auto flex max-w-4xl items-center justify-between px-6 py-3">
-          <Link href="/" className="flex items-center gap-2">
-            <Image src="/alook.svg" alt="Alook" width={22} height={22} />
-            <span className="text-lg tracking-tight" style={{ fontFamily: "var(--font-brand)", fontWeight: 700 }}>Alook</span>
+    <PublicLayout
+      maxWidth="4xl"
+      rightSlot={
+        isLoggedIn ? (
+          <Link href="/workspaces?auto" className={buttonVariants({ variant: "outline", size: "sm" })}>
+            <ArrowLeft className="mr-1.5 size-3" />
+            Back to App
           </Link>
-          <div className="flex items-center gap-2">
-            {isLoggedIn ? (
-              <Link href="/workspaces?auto" className={buttonVariants({ variant: "outline", size: "sm" })}>
-                <ArrowLeft className="mr-1.5 size-3" />
-                Back to App
-              </Link>
-            ) : (
-              <Link href="/sign-in" className={buttonVariants({ size: "sm" })}>
-                Sign In
-              </Link>
-            )}
-          </div>
-        </div>
-      </nav>
-
+        ) : (
+          <Link href="/sign-in" className={buttonVariants({ size: "sm" })}>
+            Sign In
+          </Link>
+        )
+      }
+    >
       <div className="mx-auto max-w-4xl px-6 py-10">
         {/* Breadcrumb */}
         <nav className="mb-8 flex items-center gap-1.5 text-xs text-muted-foreground">
@@ -168,6 +160,6 @@ export function TemplateDetailClient({
           </Link>
         </div>
       </div>
-    </div>
+    </PublicLayout>
   );
 }
